Refactor player components and add experience/currency
Refactored player-related components to separate identity, combat stats, experience, and currency into distinct ECS components. Player direction is now an enum (MirDirection) instead of u8, and related systems updated accordingly. Moved InputEvent enum to input.rs, and removed redundant fields from PlayerData. Updated world and map systems to use new component structure.

+++ ClientRust/src/ecs/systems/rendering/character_system.rs
@@ -115,7 +115,7 @@ impl CharacterRenderSystem {
         let current_frame = animation_tick % frame_count;
         
         // 计算最终索引：基础索引 + 方向偏移 + 帧偏移
-        action_start + (player.direction as i32 * frame_count) + current_frame
+        action_start + (player.direction as u8 as i32 * frame_count) + current_frame
     }
 
     /// 渲染单个角色
